fix: limit the delay between tasks to 60s
Ensure the checker prints some message at least once per minute,
to prevent Station Core for considering it as stuck.

As a side effect, this also fixes the problem when maxTasksPerRound
is zero, the delay is calculated as Infinity and we call
`setTimeout(Infinity)`. Such call is a no-op in Deno, which causes
the checker to loop very quickly and pollute the logs with too many
"Completed all tasks for the current round." messages.

After this change, the message is print once per minute.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>